requirements.txt
hikari-arc[rest]~=0.3.0
python-dotenv~=1.0.0
